+++ scm-webapp/src/main/java/sonia/scm/repository/DefaultRepositoryManager.java
@@ -41,6 +41,8 @@ import sonia.scm.NotFoundException;
 import sonia.scm.SCMContextProvider;
 import sonia.scm.Type;
 import sonia.scm.config.ScmConfiguration;
+import sonia.scm.event.ScmEventBus;
+import sonia.scm.security.AuthorizationChangedEvent;
 import sonia.scm.security.KeyGenerator;
 import sonia.scm.util.AssertUtil;
 import sonia.scm.util.CollectionAppender;
@@ -59,6 +61,7 @@ import java.util.Set;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ThreadFactory;
+import java.util.function.Consumer;
 import java.util.function.Predicate;
 
 import static java.util.stream.Collectors.toSet;
@@ -123,10 +126,16 @@ public class DefaultRepositoryManager extends AbstractRepositoryManager {
 
   @Override
   public Repository create(Repository repository) {
-    return create(repository, true);
+    return create(repository, r -> {
+    }, true);
   }
 
-  public Repository create(Repository repository, boolean initRepository) {
+  @Override
+  public Repository create(Repository repository, Consumer<Repository> afterCreation) {
+    return create(repository, afterCreation, true);
+  }
+
+  public Repository create(Repository repository, Consumer<Repository> afterCreation, boolean initRepository) {
     repository.setId(keyGenerator.createKey());
     repository.setNamespace(namespaceStrategyProvider.get().createNamespace(repository));
 
@@ -137,15 +146,20 @@ public class DefaultRepositoryManager extends AbstractRepositoryManager {
       RepositoryPermissions::create,
       newRepository -> fireEvent(HandlerEventType.BEFORE_CREATE, newRepository),
       newRepository -> {
-        fireEvent(HandlerEventType.CREATE, newRepository);
         if (initRepository) {
           try {
             getHandler(newRepository).create(newRepository);
-          } catch (InternalRepositoryException e) {
+            invalidateRepositoryPermissions();
+            afterCreation.accept(newRepository);
+          } catch (Exception e) {
             delete(repository);
             throw e;
           }
+        } else {
+          invalidateRepositoryPermissions();
+          afterCreation.accept(newRepository);
         }
+        fireEvent(HandlerEventType.CREATE, newRepository);
       },
       newRepository -> {
         if (repositoryDAO.contains(newRepository.getNamespaceAndName())) {
@@ -153,6 +167,10 @@ public class DefaultRepositoryManager extends AbstractRepositoryManager {
         }
       }
     );
+  }
+
+  private void invalidateRepositoryPermissions() {
+    ScmEventBus.getInstance().post(AuthorizationChangedEvent.createForEveryUser());
   }
 
   @Override
@@ -173,7 +191,8 @@ public class DefaultRepositoryManager extends AbstractRepositoryManager {
 
   @Override
   public void importRepository(Repository repository) {
-    create(repository, false);
+    create(repository, r -> {
+    }, false);
   }
 
   @Override