+++ scm-webapp/src/test/java/sonia/scm/api/v2/resources/RepositoryTypeRootResourceTest.java
@@ -27,8 +27,11 @@ package sonia.scm.api.v2.resources;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
 import com.google.inject.util.Providers;
+import org.apache.shiro.subject.Subject;
+import org.apache.shiro.util.ThreadContext;
 import org.jboss.resteasy.mock.MockHttpRequest;
 import org.jboss.resteasy.mock.MockHttpResponse;
+import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -55,10 +58,12 @@ import static org.mockito.Mockito.when;
 @RunWith(MockitoJUnitRunner.Silent.class)
 public class RepositoryTypeRootResourceTest {
 
-  private RestDispatcher dispatcher = new RestDispatcher();
+  private final RestDispatcher dispatcher = new RestDispatcher();
 
   @Mock
   private RepositoryManager repositoryManager;
+  @Mock
+  private Subject subject;
 
   private final URI baseUri = URI.create("/");
   private final ResourceLinks resourceLinks = ResourceLinksMock.createMock(baseUri);
@@ -66,13 +71,15 @@ public class RepositoryTypeRootResourceTest {
   @InjectMocks
   private RepositoryTypeToRepositoryTypeDtoMapperImpl mapper;
 
-  private List<RepositoryType> types = Lists.newArrayList(
+  private final List<RepositoryType> types = Lists.newArrayList(
     new RepositoryType("hk", "Hitchhiker", Sets.newHashSet()),
     new RepositoryType("hog", "Heart of Gold", Sets.newHashSet())
   );
 
   @Before
   public void prepareEnvironment() {
+    ThreadContext.bind(subject);
+
     when(repositoryManager.getConfiguredTypes()).thenReturn(types);
 
     RepositoryTypeCollectionToDtoMapper collectionMapper = new RepositoryTypeCollectionToDtoMapper(mapper, resourceLinks);
@@ -80,6 +87,11 @@ public class RepositoryTypeRootResourceTest {
     RepositoryTypeResource resource = new RepositoryTypeResource(repositoryManager, mapper);
     RepositoryTypeRootResource rootResource = new RepositoryTypeRootResource(Providers.of(collectionResource), Providers.of(resource));
     dispatcher.addSingletonResource(rootResource);
+  }
+
+  @After
+  public void tearDown() {
+    ThreadContext.unbindSubject();
   }
 
   @Test